split reg definitons from struct defs
