+++ src/src/scheduler.c
@@ -5,7 +5,7 @@
 #include "task.h"
 #include "def.h"
 #include "types/task-queue.h"
-#include "hardware/cortex-m3.h"
+#include "hardware/cortex-m3/context.h"
 
 #include <stdlib.h>
 #include <stdint.h>
